+++ satella/coding/decorators/__init__.py
@@ -1,9 +1,9 @@
 from .arguments import auto_adapt_to_methods, attach_arguments, for_argument, \
     execute_before
-from .decorators import wraps, chain_functions, has_keys, short_none
+from .decorators import wraps, chain_functions, has_keys, short_none, memoize
 from .preconditions import postcondition, precondition
 from .flow_control import loop_while, queue_get
 
 __all__ = ['execute_before', 'postcondition', 'precondition', 'wraps', 'queue_get', 'chain_functions',
            'has_keys', 'short_none', 'auto_adapt_to_methods', 'attach_arguments', 'for_argument',
-           'loop_while']
+           'loop_while', 'memoize']