+++ satella/coding/sequences/__init__.py
@@ -2,7 +2,7 @@ from .choose import choose, choose_one
 from .iterators import infinite_counter, take_n, is_instance, skip_first, zip_shifted, \
     stop_after, iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, \
     odd, n_th, smart_enumerate, smart_zip, unique, ConstruableIterator, walk, length, map_list, \
-    is_empty, IteratorListAdapter
+    is_empty, IteratorListAdapter, enumerate2
 from .sequences import is_last, add_next, half_cartesian, group_quantity, Multirun, make_list
 from .average import RollingArithmeticAverage
 
@@ -12,4 +12,4 @@ __all__ = ['choose', 'choose_one', 'infinite_counter', 'take_n', 'is_instance', 
            'iter_dict_of_list', 'shift', 'other_sequence_no_longer_than', 'count', 'n_th',
            'even', 'odd', 'Multirun', 'smart_enumerate', 'unique', 'map_list', 'is_empty',
            'make_list', 'RollingArithmeticAverage',
-           'IteratorListAdapter']
+           'IteratorListAdapter', 'enumerate2']