+++ tests/test_coding/test_sequences.py
@@ -4,11 +4,16 @@ import unittest
 from satella.coding.sequences import choose, choose_one, infinite_counter, take_n, is_instance, \
     is_last, add_next, half_cartesian, skip_first, zip_shifted, stop_after, group_quantity, \
     iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd, Multirun, n_th, \
-    unique, length, map_list, smart_zip, is_empty, make_list, RollingArithmeticAverage
+    unique, length, map_list, smart_zip, is_empty, make_list, RollingArithmeticAverage, \
+    enumerate2
 from satella.coding.predicates import x
 
 
 class TestSequences(unittest.TestCase):
+
+    def test_enumerate2(self):
+        a = [1, 2, 3, 4]
+        self.assertEqual(list(enumerate2(a, -1, -1)), [(-1, 1), (-2, 2), (-3, 3), (-4, 4)])
 
     def test_rolling_arith_avg(self):
         a = RollingArithmeticAverage(10)