+++ satella/coding/concurrent/__init__.py
@@ -7,9 +7,9 @@ from .thread import TerminableThread, Condition, SingleStartThread, call_in_sepa
     BogusTerminableThread
 from .id_allocator import IDAllocator
 from .timer import Timer
-from .functions import parallel_execute
+from .functions import parallel_execute, run_as_future
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
            'MonitorDict', 'MonitorList', 'Condition', 'LockedStructure', 'AtomicNumber',
            'CallNoOftenThan', 'SingleStartThread', 'IDAllocator', 'call_in_separate_thread',
-           'BogusTerminableThread', 'Timer', 'parallel_execute']
+           'BogusTerminableThread', 'Timer', 'parallel_execute', 'run_as_future']