+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.11.12
 
+* add more features to `SparseMatrix`
+