+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.11.12
 
 * add more features to `SparseMatrix`
-
+* added an extra functionality to `TimeBasedHeap`
+* added `Timer`