+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.13
+
+* change argument names for `clip`