+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.11.13
 
 * change argument names for `clip`
+* add `memoize`