+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.14
+
+* added `enumerate2`