+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.11.14
 
 * added `enumerate2`
+* added `infinite_iterator`
+* added `parallel_execute`