+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.15
+
+* added `run_as_future`