+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.11.15
-
-* added `run_as_future`
+# v2.11.16